Fixed filtering and patching

+++ src/main/java/org/n3gd0r/recipe/usecase/delete/DeleteRecipeHandler.java
@@ -3,6 +3,7 @@ package org.n3gd0r.recipe.usecase.delete;
 import org.n3gd0r.commons.mediator.HandlerFor;
 import org.n3gd0r.commons.mediator.RequestHandler;
 import org.n3gd0r.recipe.repository.RecipeRepository;
+import org.n3gd0r.recipe.usecase.exception.RequestForHandlerNullPointerException;
 import org.springframework.stereotype.Service;
 
 import lombok.extern.slf4j.Slf4j;
@@ -11,8 +12,8 @@ import lombok.extern.slf4j.Slf4j;
  * DeleteRecipe
  */
 @Slf4j
-@HandlerFor(DeleteRecipeParameters.class)
 @Service
+@HandlerFor(DeleteRecipeParameters.class)
 public class DeleteRecipeHandler implements RequestHandler<DeleteRecipeParameters, Void> {
     private final RecipeRepository repository;
 
@@ -22,6 +23,10 @@ public class DeleteRecipeHandler implements RequestHandler<DeleteRecipeParameter
 
     @Override
     public Void execute(DeleteRecipeParameters request) {
+        if (request == null) {
+            log.warn("Request DeleteRecipeParameters is null");
+            throw new RequestForHandlerNullPointerException(DeleteRecipeParameters.class);
+        }
         log.info("Deleting recipe: {}", request.id());
         repository.validateExistsById(request.id());
         repository.deleteById(request.id());
